Moved sql value for style generation
